# Speedup audmetric.detection_error_tradeoff()