As Detection Error Tradeoff is often used for pairwise similarities, the input similarity truth vectors can be very large.
By convering parts of the `detection_error_tradeoff()` function to `numpy`, a meaningful acceleration can be achieved (see below).
@hagenw 

```python
import api  # local copy of the modified repo audmetric.core.api
import audmetric
import numpy as np
import time


num_samples = np.array([1e4, 1e5, 1e6, 1e7]).astype(int)

for n in num_samples:
    # Generate seeded random truth and similarity vectors
    np.random.seed(0)
    truth = np.random.rand(n,) > 0.8
    sim = np.random.randn(n,) + truth * 1.

    # old implementation
    t_start = time.perf_counter()
    det1 = audmetric.detection_error_tradeoff(truth, sim)
    t_stop = time.perf_counter()
    rt1 = t_stop - t_start

    # new implementation
    t_start = time.perf_counter()
    det2 = api.detection_error_tradeoff(truth, sim)
    t_stop = time.perf_counter()
    rt2 = t_stop - t_start

    for a, b in zip(det1, det2):
        np.testing.assert_allclose(a, b)

    print(f"{n} samples - Runtime old: {rt1:.3f}s, new: {rt2:.3f}s; "
          f"ratio: {(rt2/rt1):.3f}")
```

```
   10000 samples - Runtime old:  0.012s, new: 0.002s; ratio: 0.188
  100000 samples - Runtime old:  0.154s, new: 0.026s; ratio: 0.169
 1000000 samples - Runtime old:  1.916s, new: 0.289s; ratio: 0.151
10000000 samples - Runtime old: 21.873s, new: 3.515s; ratio: 0.161
```