@hagenw 2024-02-28
Great, this looks indeed much faster.

What is a little bit problematic with `audmetric.detection_error_tradeoff()` is that we only have the docstring test for:

```python
>>> truth = [1, 0]
>>> prediction = [0.9, 0.1]
>>> detection_error_tradeoff(truth, prediction)
(array([1., 0.]), array([0., 0.]), array([0.1, 0.9]))
```

Would it be possible to add a few more cases to `tests/test_api.py` in a new `test_detection_error_tradeoff()` function, or would it be too complicated to get the expected values?