@maxschmitt 2024-02-28
> Would it be possible to add a few more cases to `tests/test_api.py` in a new `test_detection_error_tradeoff()` function, or would it be too complicated to get the expected values?

Should be fine, I will add some cases.