@maxschmitt 2024-02-28
`test_detection_error_tradeoff()` has been added and I also removed scalar values from `detection_error_tradeoff()` and `equal_error_rate()` inputs as these metrics are un-defined for scalar (or single-element vectors).