@hagenw 2024-02-28
As the pull request title will be used as the commit message when merging, it should only consist of 50 chars. I would propose to change the title of the pull request to: "Speedup audmetric.detection_error_tradeoff()"